refactor: modularize goalsummary component
