+++ src/components/GoalsComponents/MyGoal/GoalSummary/BudgetStartSummary.tsx
@@ -1,0 +1,21 @@
+import { GoalItem } from "@src/models/GoalItem";
+import React from "react";
+import { useTranslation } from "react-i18next";
+
+const BudgetStartSummary = ({ goal }: { goal: GoalItem }) => {
+  const { t } = useTranslation();
+  const hasStarted = goal.start && new Date(goal.start).getTime() < new Date().getTime();
+  const showStart = goal.due || !hasStarted;
+
+  if (!showStart || !goal.start) {
+    return null;
+  }
+
+  return (
+    <div>
+      {hasStarted ? t("started") : t("starts")} {new Date(goal.start).toDateString().slice(4)}
+    </div>
+  );
+};
+
+export default BudgetStartSummary;
