+++ src/components/GoalsComponents/MyGoal/MyGoal.tsx
@@ -11,7 +11,7 @@ import { darkModeState, displayPartnerMode, lastAction } from "@src/store";
 import { displayGoalId, displayUpdateGoal, goalsHistory, displayChangesModal } from "@src/store/GoalsState";
 
 import GoalAvatar from "../GoalAvatar";
-import GoalSummary from "./GoalSummary";
+import GoalSummary from "./GoalSummary/GoalSummary";
 import GoalDropdown from "./GoalDropdown";
 import GoalTitle from "./GoalTitle";
